Update script to get ignored BSPs from travis.yml
This allows IGNORED_BSPS to be configured on travis.yml to automatically
ignore BSPs known to fail on build. The list of BSPs is automatically
loaded from the "hw/bsp" so that new BSPs are already on by default and
to ignore them it is not required to change this repo (only need to add
to IGNORED_BSPS in .travis.yml from mynewt-core).