Move GH access to backend
Due to inability of forks to be able to use the Github API v3,
an extra proxy will be used, running on runkit.com.

A new XSTL to generate JSON output from RAT was added, that
enables easier parsing of data.

When parsing, new files with unknown licenses are output and
parsed/sent to runkit/GH as comments and a status is written
on the first commit with unknown files to fail the continuous
integration.

Signed-off-by: Fabio Utzig <utzig@apache.org>