Get amounts from map size; don't fail on new files
The JSON XSLT was updated to not output "unknown_amount" and
"files_amount". Those sizes are better determined by checking the len of
the dicts in Python.

The existence of new files with unknown licenses was failing the check
and was disabled for now. Now this check only fails, if there was a
communication error with the endpoints (or GH).

Signed-off-by: Fabio Utzig <utzig@apache.org>