Log added files using git parent syntax (~)
This updates the git logging command to get list of added files, using
only the commit sha, getting the parent with git syntax (~). Previously
the iteration was using the list of commits, so the parent was passed
by its own sha, which would fail when there was a single commit with added
files.

Signed-off-by: Fabio Utzig <utzig@apache.org>