Send HTTP error results to comments
Try to detect errors while posting style/license check comments and, if
errors are raised, post a new comment with the error description. This
should work for errors like https status 413, Payload too big, and maybe
a few others. Since the error might be due to instability in the service,
there wouldn't be any reason to retry, but since we don't know if that is
the case, print it to console as well and retry the POST anyway.

Signed-off-by: Fabio Utzig <utzig@apache.org>