newt_upgrade: Update some tests due to dummy mods
It's unfortunate, but some of these tests need to change any time
anything is pushed to one of the mynewt_dummy repos.  If a test leaves a
repo pointed to master (0.0.0), then the project state changes whenever
anything is pushed to the repo.

These tests should be updated to never use master / 0.0.0.  In the
meantime, this commit just updates the tests to use the latest master
branches.