Fix (hack) newt dump tests
Since stages order is now resolved using graph, the order of output is
not stable between calls, i.e. order of functions with the same stage
can be different on each call. This produces different dump output each
time "newt dump" is called so it not quite easily possible to verify
that output using diff. As a workaround, we'll compare "newt dump"
output with sysinit removed, the remaining data should be stable.